Increase LevelDB max_open_files unless on 32-bit Unix.
Summary:
```
This change significantly increases IBD performance by increasing the
amount of the UTXO index that can remain in memory. To ensure this
doesn't cause problems in the future, a static_assert on the LevelDB
version has been added, which must be updated by anyone upgrading
LevelDB.
```

Backport of core PR12495
https://github.com/bitcoin/bitcoin/pull/12495/files

Test Plan: Run IBD on CI (and expect a significant performace improvement).

Reviewers: #bitcoin_abc, deadalnix, fpelliccioni

Reviewed By: fpelliccioni

Differential Revision: https://reviews.bitcoinabc.org/D4000