[qt] Display more helpful message when adding a send address has failed
Summary:
Addresses #12796.

When we're unable to add a sending address to the address book because it
already exists as a receiving address, display a message indicating as much.
This should help avoid confusion about an address supposedly already in the
book but which isn't currently visible in the interface.

This is a partial backport of Core PR12830 : https://github.com/bitcoin/bitcoin/pull/12830/commits/8cdcaee4c7b256c5c3b70f1cfb04a5fb547311cd

Depends on D3668

Test Plan:
  make check

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3669