Don't use systray icon on inappropriate systems
Summary:
```
Prevent a user from losing access to the main window by minimizing it to
the tray on some systems (e.g. GNOME 3.26+).
```

Backport of core PR14228 and PR15000:
https://github.com/bitcoin/bitcoin/pull/14228/files
https://github.com/bitcoin/bitcoin/pull/15000/files

Test Plan:
Build and run on a Linux with a trayless DE.
Check the behavior against the summary.
Send some coins to yourself and check the notifications are popping.
Tested on Buster w/ Gnome shell.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4492