Cmake backports (#61)
* [CMAKE] Fix cached variable that prevents testing for flag support

Summary:
The `CHECK_C_COMPILER_FLAG()` and `CHECK_CXX_COMPILER_FLAG()` functions
from CMake return the result of the test in a variable passed as an
argument.
This variable is cached internally, so that subsequent calls of the
function with the same variable return the result without performing the
test again.

The `add_c_compiler_flag()` and `add_cxx_compiler_flag()` functions are
calling `CHECK_C_COMPILER_FLAG()` and `CHECK_CXX_COMPILER_FLAG()` with
the same variable `FLAG_IS_SUPPORTED`, making the test ineffective.

This diffs fixes the issue by generating a new variable name based on
the flag name. This allow for keeping advantage of the caching
mechanisme while having a different variable for each flag, making the
test working as expected.

* [CMAKE] Check for linker flag support

Summary:
CMake provides no facility to test wheter a linker flag is supported by
the compiler or not. However the CMP0056 policy (Cmake 3.2) enforces
passing the `CMAKE_EXE_LINKER_FLAGS` to the `try_compile()` function.
This offers a workaround for testing linker flag validity before adding
them, which is implemented in this diff.

Depends on D2653

* [CMAKE] Add an option to enable/disable hardening

Summary:
Tis mimics the `--enable-hardening` and `--disable-hardening` from
autotools. The default is to enable hardening.

Depends on D2653

* [CMAKE] Override the flags for the Debug CMake configuration

Summary:
This diffs overrides the default flags for the `Debug` build type,
setting the optimization to `-g3 -Og` (or their fallbacks).

Depends on D2653

* [CMAKE] Use static linkage when cross compiling for windows

Summary:
This will avoid the need to distribute dll files along with the
binaries.

Depends on D2653 and D2664

* [CMAKE] Add DEBUG and DEBUG_LOCKORDER definitions to the Debug config

Summary:
This add definitions for `DEBUG` and `DEBUG_LOCKORDER` when the Debug
build type is used.
This allow the use of `cmake -GNinja -DCMAKE_BUILD_TYPE=Debug` as an
equivalent to `configure --enable-debug`.


* [CMAKE] Harden the executables

Summary:
This adds the hardening options to cmake.
This diffs is a squashed version of D2658, D2659, D2660 plus the
hardening options for windows executables.

Depends on D2666