Avoid calling add_nodes multiple times in functional tests
Summary:
The framework is designed so that nodes are created once and can then be
started/stopped when needed. The `add_nodes` method should only be
called once when overriding `setup_nodes` or `setup_network`.
Doing otherwise causes a failure when trying to bind the newly
added node to the same address/port than a previous one after D3084 is landed.

Test Plan:
Run on Linux and OSX:
  ./test/functional/test_runner.py abc-high_priority_transaction.py

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D3090