Fix exit in generate_header.py and some formatting nits
Summary:
Possible exit solutions are `sys.exit()` or `os._exit()`, but
`os.exit()` doesn't exist. Using `sys.exit()` incidentally removes the
dependency to `os`.
Also fix some spaces in the generated code so it matches our
`clang-format` rules, as it is trivial and saves as much as 3 bytes.

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3418