Remove undue lock assertion in GuessVerificationProgess
Summary:
The `AssertLockHeld()` assumes that `cs_main` is always locked when
calling `GuessVerificationProgress()`.
However, as stated in the function comment, this might not always be
true as the lock is only required under some circumstances. The
assertion is then too restrictive.
Note that this makes our code in sync with current core codebase.

Test Plan:
With debug enabled:
  make check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4549